requirements.txt
Flask==2.2.5
Flask_Cors==4.0.0
pymongo==4.5.0
gunicorn==21.0.1
python-dotenv==1.0.0
